+++ demo/main.py
@@ -1,6 +1,8 @@
 from typing import Literal
 
-from lihil import Lihil, Route
+from lihil import Annotated, Lihil, Route, Struct, status
+from lihil.plugins.auth.jwt import JWTAuthParam
+from lihil.plugins.auth.oauth import OAuth2PasswordFlow
 
 root = Route()
 
@@ -12,18 +14,21 @@ UserNames = Literal["alice", "bob", "charlie"]
 async def get_user(user_name: UserNames | None): ...
 
 
+class UserProfile(Struct):
+    user_name: str
+    full_name: str | None = None
+    age: int | None = None
+
+
+@root.sub("/me").get(auth_scheme=OAuth2PasswordFlow(token_url="/token"))
+async def get_me(profile: Annotated[str, JWTAuthParam]) -> UserProfile:
+    return profile
+
+
+@root.sub("/status").get
+async def get_status() -> (
+    Annotated[str, status.OK] | Annotated[int, status.CREATED]
+): ...
+
+
 app = Lihil(root)
-
-
-# from fastapi import APIRouter, FastAPI
-
-
-# root = APIRouter()
-
-# @root.get("/user")
-# async def get_user(user_name: str | None):
-#     return {"user_name": user_name}
-
-
-# app = FastAPI()
-# app.include_router(root)